Session27/requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.21.0
PyYAML>=6.0
pydantic>=1.9.0 